+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added UI for creating and editing the properties of individual stimuli, including support for changing the text, image, and other attributes of the stimuli. 
 - Added parent interface to stimuli UIs to list and  allow for searching of stimuli.
 - Added a layout editor interface with its own tab.
+- Added a tab and interface for assigning trials to IAT blocks.
 ### Changed
 - Updated dependencies to latest versions
 - Modified the architecture to better support the new features and improvements, including changes to the way data is handled and processed within the application.